Models: modify migrations and model folders to change expense.amount datatype to decimal to allow deciamls

+++ migrations/20210217065830-create-expense.js
@@ -21,7 +21,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       amount: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       guestId: {
         type: Sequelize.INTEGER
